Move scene functions to own file
